feature: Aula21 - Estruturas de repetições pt05 Continue

+++ maratona-java/src/introducao/Aula06EstruturasDeRepeticoes04.java
@@ -6,7 +6,7 @@ public class Aula06EstruturasDeRepeticoes04 {
         Condição valorParcela >= 1000
     */
     public static void main(String[] args) {
-        int valorTotal = 30000;
+        double valorTotal = 30000;
         for (int parcela = 1; parcela <= valorTotal; parcela++) {
             double valorParcela = valorTotal / parcela;
             if (valorParcela >= 1000) {
